fix: small tweaks and documentation
